Use Int64 for epsilons

+++ src/bench/result.jl
@@ -39,16 +39,16 @@ end
 
 mutable struct EpsilonBenchmarkResult
     # list of all collected (non infinite) epsilons
-    epsilons::Vector{Int}
+    epsilons::Vector{Int64}
 
     # number of total samples collected
     total_samples::Int
 
     # descending list of the top worst epsilons together with the respective arguments
-    worst_arguments::TopKSortedList{Int, Tuple}
+    worst_arguments::TopKSortedList{Int64, Tuple}
 
     # the minimum epsilons of a result to be considered for the worst arguments list
-    epsilon_limit::Int
+    epsilon_limit::Int64
 
     # the call string of the function being benchmarked, with format specifiers to interpolate the values
     call_string::String
@@ -56,7 +56,7 @@ mutable struct EpsilonBenchmarkResult
     # number of collected samples that gave infinite epsilons
     no_inf_epsilons::Int
 
-    function EpsilonBenchmarkResult(call_string::AbstractString, epsilon_limit::Int, max_values::Int)
+    function EpsilonBenchmarkResult(call_string::AbstractString, epsilon_limit::Int64, max_values::Int)
         return new(
             Int[],    # epsilons vector
             0,          # total samples
@@ -68,13 +68,13 @@ mutable struct EpsilonBenchmarkResult
     end
 end
 
-function Base.insert!(eps::EpsilonBenchmarkResult, key::Int, value::Tuple)
+function Base.insert!(eps::EpsilonBenchmarkResult, key::Int64, value::Tuple)
     eps.total_samples += 1
 
     if key >= eps.epsilon_limit
         insert!(eps.worst_arguments, key, value)
     end
-    if key != typemax(Int)
+    if key != typemax(Int64)
         push!(eps.epsilons, key)
     else
         eps.no_inf_epsilons += 1
