+++ test/utils.jl
@@ -9,7 +9,7 @@ FLOAT_TYPES = [Float16, Float32, Float64]
         # As a user, do *not* use this constructor!
         p = PrecisionCarrier(one(F), big(1.0 + 1.0e-5))
 
-        @test epsilons(p) == round(Int, 1.0e-5 / eps(F))
+        @test epsilons(p) == round(Int64, 1.0e-5 / eps(F))
 
         @test epsilons(precify(F(Inf))) == 0
         @test epsilons(precify(F(-Inf))) == 0
@@ -17,16 +17,16 @@ FLOAT_TYPES = [Float16, Float32, Float64]
 
         # test some edge cases
         @test epsilons(PrecisionCarrier{F}(0, big(0.0))) == 0
-        @test epsilons(PrecisionCarrier{F}(0, big(1.0))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(Inf, big(1.0))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(-Inf, big(1.0))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(NaN, big(1.0))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(Inf, big(-Inf))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(-Inf, big(Inf))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(Inf, big(NaN))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(-Inf, big(NaN))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(NaN, big(Inf))) == typemax(Int)
-        @test epsilons(PrecisionCarrier{F}(NaN, big(-Inf))) == typemax(Int)
+        @test epsilons(PrecisionCarrier{F}(0, big(1.0))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(Inf, big(1.0))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(-Inf, big(1.0))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(NaN, big(1.0))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(Inf, big(-Inf))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(-Inf, big(Inf))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(Inf, big(NaN))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(-Inf, big(NaN))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(NaN, big(Inf))) == typemax(Int64)
+        @test epsilons(PrecisionCarrier{F}(NaN, big(-Inf))) == typemax(Int64)
     end
 
     @testset "significant_digits" begin
